+++ docs/csharp/programming-guide/classes-and-structs/snippets/properties/interfaces.cs
