+++ samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideIndexers/CS/Indexers.cs
@@ -109,163 +109,6 @@ class Program
 }
 //</Snippet2>
 
-namespace Wrap2
-{
-    //<Snippet5>
-    public class BaseClass
-    {
-        private string _name = "Name-BaseClass";
-        private string _id = "ID-BaseClass";
-
-        public string Name
-        {
-            get { return _name; }
-            set { }
-        }
-
-        public string Id
-        {
-            get { return _id; }
-            set { }
-        }
-    }
-
-    public class DerivedClass : BaseClass
-    {
-        private string _name = "Name-DerivedClass";
-        private string _id = "ID-DerivedClass";
-
-        new public string Name
-        {
-            get
-            {
-                return _name;
-            }
-
-            // Using "protected" would make the set accessor not accessible.
-            set
-            {
-                _name = value;
-            }
-        }
-
-        // Using private on the following property hides it in the Main Class.
-        // Any assignment to the property will use Id in BaseClass.
-        new private string Id
-        {
-            get
-            {
-                return _id;
-            }
-            set
-            {
-                _id = value;
-            }
-        }
-    }
-
-    class MainClass
-    {
-        static void Main()
-        {
-            BaseClass b1 = new BaseClass();
-            DerivedClass d1 = new DerivedClass();
-
-            b1.Name = "Mary";
-            d1.Name = "John";
-
-            b1.Id = "Mary123";
-            d1.Id = "John123";  // The BaseClass.Id property is called.
-
-            System.Console.WriteLine("Base: {0}, {1}", b1.Name, b1.Id);
-            System.Console.WriteLine("Derived: {0}, {1}", d1.Name, d1.Id);
-
-            // Keep the console window open in debug mode.
-            System.Console.WriteLine("Press any key to exit.");
-            System.Console.ReadKey();
-        }
-    }
-    /* Output:
-        Base: Name-BaseClass, ID-BaseClass
-        Derived: John, ID-BaseClass
-    */
-    //</Snippet5>
-}
-
-class Test
-{
-
-    //<Snippet6>
-    private string _name = "Hello";
-
-    public string Name
-    {
-        get
-        {
-            return _name;
-        }
-        protected set
-        {
-            _name = value;
-        }
-    }
-    //</Snippet6>
-}
-
-//<Snippet7>
-public class Parent
-{
-    public virtual int TestProperty
-    {
-        // Notice the accessor accessibility level.
-        protected set { }
-
-        // No access modifier is used here.
-        get { return 0; }
-    }
-}
-public class Kid : Parent
-{
-    public override int TestProperty
-    {
-        // Use the same accessibility level as in the overridden accessor.
-        protected set { }
-
-        // Cannot use access modifier here.
-        get { return 0; }
-    }
-}
-//</Snippet7>
-
-namespace WrapISomeInterface
-{
-    //<Snippet8>
-    public interface ISomeInterface
-    {
-        int TestProperty
-        {
-            // No access modifier allowed here
-            // because this is an interface.
-            get;
-        }
-    }
-
-    public class TestClass : ISomeInterface
-    {
-        public int TestProperty
-        {
-            // Cannot use access modifier here because
-            // this is an interface implementation.
-            get { return 10; }
-
-            // Interface property does not have set accessor,
-            // so access modifier is allowed.
-            protected set { }
-        }
-    }
-    //</Snippet8>
-}
-
 namespace WrapProgram
 {
     //<Snippet9>
